Create Readlines_rstrip_txtfile
Different from the simple open and read with read(), this uses readlines() which reads all lines of the txt file and returns them as a list of strings.
However it displays new lines as \n so we have to get rid of these with the rstrip() (right hand side strip) method. 
Also the r in the absolute file path allows us to use the absolute path without having to escape the backslashes.
I use a while loop for this.

note Be careful that you are using the correct variable names in your while loop.